requirements.txt
alembic==1.16.5
altgraph==0.17.4
art==6.3
attrs==25.3.0
beautifulsoup4==4.14.2
blinker==1.9.0
cachetools==6.1.0
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
config==0.5.1
cowsay==6.1
DateTime==5.5
downloads==1.1.0
Flask==3.1.2
Flask-CKEditor==1.0.0
flask-cors==6.0.1
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.2.4
h11==0.16.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
Nuitka==2.7.12
null==0.6.1
ordered-set==4.1.0
outcome==1.3.0.post0
packaging==24.2
pefile==2023.2.7
pillow==11.2.1
py2exe==0.14.0.0
pycparser==2.22
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2024.10
PySocks==1.7.1
python-dotenv==1.1.1
pytz==2025.2
pywin32-ctypes==0.2.3
requests==2.32.4
secret-key-generator==0.0.8
selenium==4.34.2
setuptools==75.5.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
SQLAlchemy==2.0.43
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.15.0
urllib3==2.5.0
webdriver-manager==4.0.2
websocket-client==1.8.0
Werkzeug==3.1.3
wheel==0.45.0
wsproto==1.2.0
WTForms==3.2.1
zope.interface==8.0.1
zstandard==0.23.0
